reuse gesture code for attractor line demo
